requirements.txt
asgiref==3.8.1
async-generator==1.10
attrs==23.2.0
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.16.0
colorama==0.4.6
coverage==7.6.0
cryptography==43.0.0
Django==5.0.7
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
exceptiongroup==1.2.2
Faker==26.0.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
mccabe==0.7.0
mypy==1.11.1
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.1
parameterized==0.9.0
paramiko==2.11.0
pillow==10.4.0
pluggy==1.5.0
py==1.11.0
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.2
pytest-django==4.8.0
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
toml==0.10.2
tomli==2.0.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
